requirements.txt
Flask==3.0.0
fpdf==1.7.2
Werkzeug==3.0.1
gunicorn==21.2.0
Pillow==10.0.0

